+++ app/src/main/java/com/chrisvasqm/cuadramo/catalog/CatalogAdapter.kt
@@ -8,7 +8,7 @@ import androidx.fragment.app.FragmentManager
 import androidx.recyclerview.widget.RecyclerView
 import com.chrisvasqm.cuadramo.R
 import com.chrisvasqm.cuadramo.model.Cuadre
-import com.chrisvasqm.cuadramo.ui.ItemOptionsBottomSheetDialogFragment
+import com.chrisvasqm.cuadramo.dialogs.ItemOptionsBottomSheetDialogFragment
 
 class CatalogAdapter(private val cuadres: MutableList<Cuadre>, private val manager: FragmentManager) : RecyclerView.Adapter<CatalogAdapter.ViewHolder>() {
