+++ app/src/main/java/com/chrisvasqm/cuadramo/editor/EditorActivity.kt
@@ -13,7 +13,7 @@ import com.chrisvasqm.cuadramo.model.Cuadre
 import com.chrisvasqm.cuadramo.databinding.ActivityEditorBinding
 import com.chrisvasqm.cuadramo.extensions.clear
 import com.chrisvasqm.cuadramo.extensions.toInt
-import com.chrisvasqm.cuadramo.ui.PreviewBottomSheetDialogFragment
+import com.chrisvasqm.cuadramo.dialogs.PreviewBottomSheetDialogFragment
 import com.google.android.material.snackbar.Snackbar
 import com.google.android.material.textfield.TextInputEditText
 import com.google.android.material.textfield.TextInputLayout
